+++ src/Umbraco.Core/Configuration/IGlobalSettings.cs
@@ -58,15 +58,6 @@
         bool UseHttps { get; }
 
         /// <summary>
-        /// Gets a value indicating whether the content dashboard should be available to all users.
-        /// </summary>
-        /// <value>
-        ///     <c>true</c> if the dashboard is visible for all user groups; otherwise, <c>false</c>
-        ///     and the default access rules for that dashboard will be in use.
-        /// </value>
-        bool AllowContentDashboardAccessToAllUsers { get; }
-
-        /// <summary>
         /// Returns a string value to determine if umbraco should skip version-checking.
         /// </summary>
         /// <value>The version check period in days (0 = never).</value>
