+++ src/Umbraco.Core/Dashboards/IContentDashboardSettings.cs
@@ -1,13 +1,14 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Umbraco.Core.Dashboards
+﻿namespace Umbraco.Core.Dashboards
 {
     public interface IContentDashboardSettings
     {
-        IAccessRule[] GetAccessRulesFromConfig();
+        /// <summary>
+        /// Gets a value indicating whether the content dashboard should be available to all users.
+        /// </summary>
+        /// <value>
+        ///     <c>true</c> if the dashboard is visible for all user groups; otherwise, <c>false</c>
+        ///     and the default access rules for that dashboard will be in use.
+        /// </value>
+        bool AllowContentDashboardAccessToAllUsers { get; }
     }
 }
